Finished corrections based on feedback
